Changed from separate 2d and 3d point classes to one 3d point class.
Instead of trying to keep track of which points are 2d and which are 3d I am
just going to make everything work with 3d points. It is a more general form of
a point.

I think I can make this work for the section property calculator as well.

All tests pass