Restructured section property class.
Removed all the class members. The user can directly request the section
properties they want and the class run those methods.

Now that I am thinking about this I wonder if there is a way to save some
computation time by calculating the property the first time the user requests it
then storing it. I might also be getting to complicated here.

I mostly ran into a problem trying to test the class. If everything is run on
initializing the class how can one easily test each function of the class. To be
further considered.

It would be nice if everything was calculated when the class was initialized and
then just grab the data members instead of having to calculate the properties
each time they are requested. However, the user may not need all the properties
so there will be wasted computation time.