Testing section property class.
Having trouble getting the functools.cmp_to_key function working with the less
function that compares the points.

I think I am going to have to play with it some more. I currently have some
print statements in the section_property class to see how it is sorting points.

The unittesting for this is less than helpful for actually solving the problem.
I am going to have to walk through the logic to try to understand.

This stackoverflow thread may be useful:
https://stackoverflow.com/questions/32752739/python-how-does-the-functools-cmp-to-key-function-works

I think this is the lynch-pin to the whole section property class working fairly
flawlessly. However I have said that before about other software (mlob) that
took weeks to figure out and get right.