Progress on section property class. Sorting/comparison functionality up.
The sorting/comparison functions are up and running now. They actually attempt
to sort the points. I need to walk through a basic example and tweak the logic
in the comparison function to correctly sort the points.

Added a start_pt function. Needed to find the lower left most point of the shape
to start sorting from there. The original algorithm started sorting from the 12
o'clock position but if I can get it to work from a different point I think it
will be cleaner.

Still working on it. Need to update testing routines.