Continuing to figure out the how the algorithm works.
It works as expected. It sorts points from 12 o'clock ccw. I tried to have it
sort with the lower left most point but it breaks. I need to think about it some
more. I think either it has to be a point inside the polygon or I might be able
to take the start point out of the list and sort about that point then re-insert
it at the beginning and the end so it closes.

Additional investigations and work to be complete.