Sorting points works. Continued to write testing.
Decided to leave the sorting alone for now. It sorts from the 12 o'clock
position and it seems to work fine with the algorithm to calculate section
properties. I may come back to this later but for now I don't know that it
matters.

Testing other functions with a simple test at first. I plan to write more
complex tests in the future.

I have also thought about some problems that need to be overcome.
1. If a user wants to just enter the section property information how is that
handled?
2. Checks to make sure points are in the first quandrant

It's a work in progress.