Reorganized analysis, member.py to element.py, reorganized geometry
I think the base object should be an element from there one can have points or lines or surfaces. 

I think the geometry should be exactly that, geometry. Points, lines, and surfaces (eventually, maybe). The element defines the geometry type and stiffness matrix. Geometry could also be nodes, members, and planes or something but to keep is simple and generic. 

The utils module is dwindling. Really the matrix decomposition is also going to go away because the LU decomposition is going in the naf module outside of the fsap module. I am going to leave the utils module because I will probably find some use for it. 

Reorganized the analysis module. I am not sure if those stiffness matrices will stay in the analysis module or if the analysis module will go away. I guess a class for each one and assign it like a property, similar to a material or section property, to the element. Yea, I like that idea.