Writing functions to read input file.
Noticed that these functions will need to be updated to read in different
structure types (truss, 2D frame, 3D frame) because the section properties
information required changes.

Also for a 3D frame the joint location function will have to change to read in
an additional coordinate (z).

For now I am going to continue to write this as simple as possible. The question
because do I have a different input file reader for each different type of input
file or do I have on input file reader that handles everything?