Ran a basic test and everything seems to work. Made changes need to retest.
Changed the input file to include joint numbers instead of joint 1 being in the
first position. It just provides more flexibility to the user.

Moved the sample/test input file from test into the input folder. I'm not sure
if this will be the final location.

Wrote documentation for the input file so I can remember what all the cryptic
numbers stand for.

The output file should echo all the input information.