Started exploring object oriented aspects of fsap.
It really baloons the code base quickly.

I think there are some advantages. I think it is easier to think about it terms
of objects. Not totally sure how it all links together. Not sure how the
material property class makes a lot of sense but it might for some non-linear
crap.

Anyway didn't want to muddy the code base too much. I still don't have anything
that is even close to working and I didn't want this huge fracture in the code
base.