Started working on function to assemble structure stiffness matrix.
Spent a lot time this evening playing with the object oriented parts of fsap.

I would like to stay focused on the example I am working from in Kassimali even
though it is a more procedural approach. First I need to understand in detail
how it works then I can build it out.

Saved the oop work in a seperate branch. I think I was on the right track just
not where I wanted to be.

Bed time...