Still working on assembling stiffness matrix.
Ran into a problem with my implementation scheme. If I want to let the user name
the joints then I have to look up the joint name to get the joint coordinates.
This doesn't work very well as a list. Had to change the data structure in the
ri.py module to be a dictonary for 'jt'.

Work in progress.