Change to the input file format.
It was proving to difficult to write the program laid out in Kassamali and make
my own personal additions to it. Like being able to "name" the joints and
members directly in the input file.

I do think this implementation is possible and would be of great benefit to the
user. However, on my first pass through the example program from Kassamali it is
taking too much effort to implement the changes.