Added to the documentation.
I am planning for the documentation of this example to over the top. I want
it to be very detailed so that I have to think about exactly what the
program is doing and why, and document it so I can return to my
understanding later.

This will serve when I embark I writing my version of it and how it all
has to fit together.

I will also probably be able to eventually borrow most of this documentation
for my version of the program. Wa-hoo, documentation already done! I just
get to write the code.