Added longer comments to the scv function.
The scv function generates the structure coordinate vector which is the
lynch pin to assemblying the structure stiffness matrix programmatically.
It keeps track of the restrained and unrestrained structure DOFs with
respect to each joint of the structure. This enables a look-up function to
determine based on the structure connectivity which member stiffnesses
should be added together in the structure stiffness matrix.