Started debugging and getting things to run.
The full program is not complete there is still some work to be done on getting
member local forces and deflections but it is far enough along to test.

Had to 'install' naf package with the linalg module for the LU decomposition
function. See the following links for an explanation:
https://pip.pypa.io/en/stable/reference/pip_install/#editable-installs
https://python-packaging.readthedocs.io/en/latest/minimal.html

I would be nice if fsap automatically pulled those repos when it is installed
on a new computer or when the github repo is cloned. This is beyond my
understanding and knowledge at this point.

Testing the first bit of the example. :)