We have correct output for one example problem.
Holy shit. Fuck this example. The off by one errors are a fucking mess. Like
does the index start at 1? or 0? The joint numbering makes it even more of a
mess and it tries to use the joint numbering through out. Why not make life
easier and just have a dictonary that "converts" the joint numbers to indexes
since that is essentially what we are doing all over the place. Just simplify
it.

But hey I am learning something and with the off by one errors I am actually
having to think harder about what the program is doing so I can fix it. That is
probably a good thing so I actually understand what is going on when I start to
write my own version.

We have output for one example. Holy shit! :)