Add new function to get member properties, add skeleton for last part of program
The function to get member properties is repeated later in the program. Even
though it is repeated in the example I just couldn't do it. It's not good
programming practice.

Added skeleton for the last part of the program, calculating the member end
displacements and forces, and support reactions.

Gotta get back to the work that pays the bills.