Pushed results from model to rse. Rse to call output module.
I would like allow the user to be able to set the level of verbosity for the
output. I mean it is all there and it is fairly trivial to print it all to a
screen. For now though keeping it simple.

Gathered up results from model and passed them to rse.

Need to write output module. Will echo input and output all in one. I think I
will do both, write it to the screen and to a file.

Good progress.