Outlined joint and element classes. To lazy to get notes and do real work.
Outlined the base and sub classes for the joint and element classes. I actually
think the joint classes are done. I need to look at my notes/book for some of
the algorithems and I was too lazy to get of the couch to get those things. It's
easier to do the boring stuff when you are kinda tired.